Rename io.py to reporters.py (#29)
* Fix Typo in README and cli.py

* Rename io.py to reporters.py
